--- src/Generator.java
@@ -1,3 +1,0 @@
-public interface Generator {
-    public void generateMaze(Maze maze);
-}
