+++ src/entities/MazeIterator.java
